Alligator Food Choice

Agresti (1996, p. 207) gives this data on 59 alligators sampled from a lake in Florida. It has the length of the alligator in meters and the primary food type found in the alligator's stomach. The food type was classified into three categories: "Fish", "Invertebrates", and "Other".

Of interest is whether or not the length of an alligator is associated with the primary food type. Does knowing the length of an alligator give us some indication about its primary food type? If so, how is length associated with the choice of food type?

Usage

data("gators", package = "nestedLogit")

Format

A data frame with 59 rows and 2 columns.

food

Primary food type found in the alligator's stomach, a factor with levels "Other", "Fish", and "Invertebrates".

length

Length of the alligator in meters, a numeric vector.

Source

Agresti, A. (1996). An Introduction to Categorical Data Analysis. Wiley.

References

An example using this from https://data.library.virginia.edu/getting-started-with-multinomial-logit-models/.

See also

Examples

data(gators)
#> Warning: data set 'gators' not found
table(gators$food)
#> Error: object 'gators' not found
# average length of gators by food
with(gators, tapply(length, food, mean))
#> Error: object 'gators' not found

# create dichotomies
dichot <- logits(d1=dichotomy("Other", c("Fish", "Invertebrates")),
                 d2=dichotomy("Fish", "Invertebrates"))

gators.nested <- nestedLogit(food ~ length,
                            dichotomies = dichot,
                            data = gators)
#> Error: object 'gators' not found
car::Anova(gators.nested)
#> Error: object 'gators.nested' not found

# use the plot method
plot(gators.nested, x.var = "length")
#> Error: object 'gators.nested' not found