Convert a predictNestedLogit object to a data.frame

Convert a predictNestedLogit object to a data.frame

Usage

# S3 method for predictNestedLogit
as.data.frame(x, row.names = NULL, optional = FALSE, newdata, ...)

Arguments

x

a "predictNestedLogit" object

row.names

row.names for result (for conformity with generic; not currently used)

optional

logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional

newdata

A newdata data.frame used to generate predicted values. If not supplied, the original data frame is used.

...

other arguments (unused)

Value

A data frame containing the newdata values of predictors along with the columns response, p, se.p, logit, se.logit

Examples

data("Womenlf", package = "carData")
comparisons <- logits(work=dichotomy("not.work", c("parttime", "fulltime")),
                     full=dichotomy("parttime", "fulltime"))

wlf.nested <- nestedLogit(partic ~ hincome + children,
                          dichotomies = comparisons,
                          data=Womenlf)
# get predicted values for a grid of `hincome` and `children`
new <- expand.grid(hincome=seq(0, 45, length=10),
                   children=c("absent", "present"))

pred.nested <- predict(wlf.nested, new)
plotdata <- as.data.frame(pred.nested, newdata=new)
str(plotdata)
#> 'data.frame':	60 obs. of  7 variables:
#>  $ hincome : num  0 0 0 5 5 5 10 10 10 15 ...
#>  $ children: Factor w/ 2 levels "absent","present": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ response: chr  "not.work" "parttime" "fulltime" "not.work" ...
#>  $ p       : num  0.2082 0.0237 0.7681 0.2452 0.0378 ...
#>  $ se.p    : num  0.0633 0.0177 0.0639 0.058 0.0222 ...
#>  $ logit   : num  -1.34 -3.72 1.2 -1.12 -3.24 ...
#>  $ se.logit: num  0.384 0.767 0.358 0.313 0.61 ...